Added booster command for player Added boosters loading from database

+++ src/main/java/pl/timsixth/boostersaddon/model/TemporaryBooster.java
@@ -4,7 +4,7 @@ import java.util.concurrent.TimeUnit;
 
 public interface TemporaryBooster extends Booster {
 
-    long getTime();
+    int getTime();
 
     TimeUnit getTimeUnit();
 }
